Rename files to avoid confusion.
Send-email, send-qr-email, send-task-sheet was getting confusing.

Related #54
