+++ src/templates/views/qr-email.hbs
