+++ src/templates/views/task-sheet-email.hbs
@@ -1,3 +1,3 @@
 <h2>
-  Hello {{ firstName }} You can find your task sheet <a href="{{ taskSheetURL }}">here!</a>
+  Hello {{ firstName }}, you can find your task sheet <a href="{{ taskSheetURL }}">here!</a>
 </h2>
