+++ src/templates/views/task-sheet.hbs
@@ -1,2 +1,1 @@
 <h1>TASK SHEET HERE</h1>
-<h2>HELLO {{firstName}} {{lastName}}!</h2>
